+++ NEWS.md
@@ -10,6 +10,11 @@ qq, quantile and probability plots to a single graph.
 3. Fix #126 - Add function `util_uniform_param_estimate()`
 4. Fix #127 - Add function `util_cauchy_param_estimate()`
 5. Fix #130 - Add function `tidy_t()` - Also add to plotting functions.
+6. Fix #151 - Add function `tidy_mixture_density()`
+7. Fix #150 - Add function `util_geometric_stats_tbl()`
+8. Fix #149 - Add function `util_hypergeometric_stats_tbl()`
+9. Fix #148 - Add function `util_logistic_stats_tbl()`
+10. Fix #147 - Add function `util_lognormal_stats_tbl()`
 
 ## Minor Fixes and Improvements
 1. Fix #110 - Bug fix, correct the `p` calculation in `tidy_poisson()` that will